Implement a prototype strip encoder
The idea is as follows:

We first use our stripifier to produce a reasonably short sequence of
strips with restarts & swaps (see below). This is done after optimizing
for vertex cache - which means that ACMR of the strip sequence is
reasonably good, ~5% worse than triangle list.

After this we optimize for vertex cache, which reorders the vertices to
be sequential. Now the fun begins.

Whenever you output a triangle in the strip sequence we generate, you
get one of the following cases:

- Strip start; encoded using a restart index (-1) and three vertex
indices, a b c
- Strip cont(inuation); encoded using a vertex index
- Strip swap; this is a special construct that flips the strip so that
the next outgoing edge is on the opposite side of the "normal" strip
order. This is encoded like this: instead of "a b c" you get "a b a c"
which outputs the same triangle as "a b c" would, but the next vertex
after this will use the edge ac instead of bc.

The extra vertex in strip cont & strip swap cases is very often
predictable perfectly - this is because we reordered vertices to be
sequential. We thus use the following 2-bit triangle code:

00 - cont with vertex=next; no extra data (outputs 1 triangle)
01 - swap with vertex=next; no extra data (outputs 1 triangle)
10 - cont/swap with vertex encoded using extra data
11 - restart with 3 vertices encoded using extra data

For cont/swap we need to encode the vertex index *and* 1 bit that
differentiates between the two.
For restart we need to encode three new triangle vertex indices.

Whenever we encode vertex indices explicitly, we use a 4-bit buffer (to
simplify bit I/O), where we encode the distance between next (high
watermark encoding) and vertex. If it doesn't fit into 4 bits, we encode
it using a pseudo varint encoding.

The details of the 4-bit encoding need a lot of work, I came up with
this as I was writing code on stream...

The good news! This code is super simple. It also seems reasonably
efficient.
The bad news! After gzip, we get substantially worse results.

IdxCodec : 9.0 bits/triangle (post-deflate 4.9 bits/triangle)
StrpCodec: 7.4 bits/triangle (post-deflate 6.3 bits/triangle)

I was hoping for bits/triangle around 5-6 in the uncompressed case, but
it didn't happen. I need to study the distribution of the values that
I'm encoding; it also looks like there's some suspicious strip sequences
where swaps happen way too often. I'm not sure if this is inevitable or
if this means that I can improve the stripifier to produce better input
data for this codec.

(note that the stripifier *very* intentionally tries to optimize for
ACMR at the expense of strip length, so non-optimal strips are kinda a
given, but I expected better).

[Update]

If we enable the VByte variant that relies more on zlib compression, we
get this:

StrpCodec: 8.7 bits/triangle (post-deflate 6.2 bits/triangle)

This suggests that the 4-bit coding is actually reasonably meaningful.
In vbyte mode we'd expect most of the stream - 6.7 bits/triangle avg -
to go through zlib entropy coding, so the fact that the naive 4-bit
coding gets just 0.1 bits/triangle extra means that it's going to be
hard to improve on 4-bit coding unless we find other ways to
predict the data better.